plugin/doc-reader/requirements.txt
dify_plugin>=0.0.1b44
llama-index-readers-file
llama-index-readers-minio
docx2txt
platformdirs